fix: data stores not updating on save import
